+++ src/Kafka/Common/Errors/GroupCoordinatorNotAvailableException.php
@@ -18,7 +18,7 @@ use Exception;
 /**
  * The group coordinator is not available.
  *
- * Named ConsumerCoordinatorNotAvailableCode (15) in kafka/common/ErrorMapping.scala @ 0.8.2.2.
+ * Named ConsumerCoordinatorNotAvailableCode (15) in kafka/common/ErrorMapping.scala @ 0.9.0.1.
  */
 class GroupCoordinatorNotAvailableException extends KafkaException implements RetriableException, ServerExceptionInterface
 {